Hoja seleccionable, scrolleable y editable. Tremendo

+++ app/src/main/java/my/life/tracker/agenda/interfaces/ActividadDao.kt
@@ -11,16 +11,16 @@ interface ActividadDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     fun addActividad(actividad: Actividad) : Long
     @Query("SELECT * FROM Actividad WHERE id = :id")
-    fun getActividadById(id:Int): Actividad?
+    fun getActividadById(id:Long): Actividad?
 
     @Delete
     fun deleteActividad(actividad: Actividad)
     @Query("delete from Actividad where id in (:idList)")
-    fun deleteActividades(idList: List<Int>)
+    fun deleteActividades(idList: List<Long>)
 
     @Query("SELECT * FROM Actividad")
     fun getActividades() : List<Actividad>
 
     @Query("SELECT * FROM Actividad WHERE Actividad.id IN (:actividades)")
-    fun getActividadesByIdList(actividades:List<Int>) : List<Actividad>
+    fun getActividadesByIdList(actividades:List<Long>) : List<Actividad>
 }
